Moving table components into component directory
